Gracefully handle manuals that can't be found
Pattern copied from app/repositories/specialist_document_repository.rb and
app/controllers/abstract_documents_controller.rb.

If you navigate to the path for a manual that doesn't exist in the database, you
currently see a generic Rails error page. This happens routinely when a manual
is created in our Preview environment, gets lost when the production => preview
data sync happens and then a user tries to access it the following day. By
raising a more specific error, we can then catch it and handle it gracefully.