Deduplicate Manual Service Registries' common bits
These registries were providing an almost identical set of services, with each
shared service duplicating the exact same setup requirements.

The difference between the two types of registry is the repository they use.
The ManualServiceRegistry uses a privileged access-anything-it-likes
repository, whereas the OrganisationalManualServiceRegistry uses a repository
that's restricted by organisation.

It's hard to know when's the right time to use class inheritance. There are
only a few uses of it in Specialist Publisher. An alternative approach might
look like

  ManualServiceRegistry.new(
    repository: privileged_repository,
    ...
  )

and

  organisation_repository = repository_for_organisation(my_organisation_slug)

  ManualServiceRegistry.new(
    repository: organisation_repository,
    ...
  )

every time we need the service registry.

Or, the "config" (of injecting the correct repository) could be setup for us in
a method publicly available in a registry of its own. RegistryRegistry, anyone?

Most other options seemed to require that one repository was the default, and
the other was treat as an override of that default - e.g. in the case of
"decorating" one registry to behave in a more or less privileged manner.

So I've chosen to hide each registry's "config" inside its own class, with the
"organisational" class handling the task of taking an `organisation_slug` and
finding the right repository.