Modify GDS editor test to check for org slug
After the previous commit, we can modify the test which checked that when
a GDS editor modified a manual created by a user from another Org, it
kept the previous Orgs slug. This test now visits the Manual page and
checks that the slug is rendered on the show page.