Removing content domain initializer
This file was previously matching the environment and returning the dev
application that the format could be ran on. With this change we
directly use Plek to return the proper environment root URL in order to
direct the Editor to the correct site.

The previous behaviour was also incorrectly using http for Preview and
pointing Editors on Staging to Production.