Access attributes correctly on wrapped class.
The document passed to the formatter is not directly an instance of
SpecialistDocument, it is wrapped by several other classes including the
specific model for the document_type. This wrapping model removes the
`extra_fields` data from the `attributes` hash and makes its members
available directly as attributes.

Updated the formatter to give access to the basic rendered document,
rather than just the attributes hash, and updated the test so that it
passes in the wrapped class.