Send correct format for ESIF documents to email-alert-api
We were sending "esi_fund" to email-alert-api because that's the string used as
the "document_type" internal to specialist-publisher. All of the other
exporters, however, use "european_structural_investment_fund". This means that
email subscriptions are created for the format
"european_structural_investment_fund", but email alerts are triggered for
"esi_fund", therefore, there aren't any matching subscriptions and so emails
weren't sent.