Migration to export all specialist documents.
Sends everything to publishing API. For published editions, uses an
overridden format, "placeholder_specialist_document", so that it does
not incur the overhead of re-registering existing routes with the
router.

This does not affect the data itself, or the front-end representation.
The format name can be correct directly in content-store afterwards.