Clarify manual creation steps
Prior to 5f03c5598cedc we did not create documents during the 'Given a
draft manual exists' step. That commit needed a manual with some
documents to test the speed up in manual publishing.

However, having a manual created always with some documents makes it
harder to test sending draft manual documents to the draft publishing
api.

So, I'm rewording all of these steps to be explicit about whether the
draft manual does or does not have documents.