Add topic details to some finders.
These are the topics that are currently set in panopticon.

Specialist-publisher doesn't send finders to panopticon/contentapi
(finder-api used to but is no longer in use), so the topic tags can no
longer be managed there. There's also no way for tagging information
specified in panopticon to make it into rummager.

Specifying these details here will allow the topics to be sent to
rummager.