Only mark a manual document exported when sending to live
The 'marking as exported' is only relevant to the sending of the
document to the live publishing api. We don't want to mark a document as
exported when it's sent the the draft API. If we did so, we'd prevent
the new document from being sent to the live API when needed.

This includes an extra step in the publishing-a-manual cucumber feature
to verify that the manual document is sent to the draft endpoint when
editing an already published manual.