Move organisation_api into wiring
since it's used in several places.

Also define a caching organisation_fetcher service.

I wondered whether the caching is necessary in fact, this was added as
part of the 'speeding up manuals publishing' task, but now that the
organisation_api is a singleton, we can benefit from the LRU cache
provided by the gds api adapters.