Add leading slash to slugs when sending to Rummager
The id and link fields sent to Rummager must include the leading slash in the
path. The link field is used for links to the content item on pages which pull
their content from Rummager, for example search results and topic pages (as
recently changed in the Collections app).

When the link field doesn't start with a slash, the link is treated as
relative to the current URL. In search results this means that the links are
right by coincidence: the current URL has a single path segment
(/search?q=foo) so the link is interpreted relative to / which happens to be
correct. However on topic pages, which have more path segments, the links are
wrong, for example on the topic page

    /topic/government-digital-guidance/content-publishing

the link to the Style Guide is wrongly shown as

    /topic/government-digital-guidance/guidance/style-guide

instead of

    /guidance/style-guide

Including the leading slash in the fields for Rummager fixes these links.