Refactor ManualSectionIndexableFormatter to use abstract class
This avoids duplication of with_leading_slash, and makes this formatter more
consistent with the others.

The main change here is to override private methods from the superclass
instead of defining the attributes entirely in indexable_attributes. This
makes it clearer that the only attribute unique to this formatter is the
manual.