Add script to fix id and link in existing documents in Rummager
We need to migrate the existing documents in Rummager to include the leading
slash in the id and link fields. Because the id is changing, we need to
republish and then delete the document with the old id. This should be done
for each document individually (rather than in batches) to minimise the time
during which each document will be in the search index twice.

This is heavily based on the existing scripts to republish manuals and
specialist documents.

For manuals, republishing will send them to the publishing API as well as
Rummager, but for specialist documents republishing only sends to Rummager. In
this case that inconsistency doesn't matter because we've only changed the
Rummager data format.

Rummager returns 200 for a request to delete a document which doesn't exist,
so we don't need to check that each one exists before deleting it.