Handle tags not always being present on a manual
Even though all already-published manuals in Production with tags in
Panopticon have been published again to pick up the tags to send to Rummager,
some published manuals in a database replicated from Preview 2 days ago didn't
have the tags populated in Specialist Publisher. Tags aren't fetched when
republishing, so we need to handle the tags not being present when
republishing in order for the script to fix leading slashes to run without
errors.