Show preview link when manual's document is draft
Previously the preview link would only display when the manual itself
was edited; an edit to a document belonging to that manual does not
change the draft status of the manual itself so the preview link would
not be displayed. Changed the logic so that it also checks the child
documents' statuses.