Fix script to correctly delete old ESIF document
The script to fix the missing leading slash on the id and link fields sent to
Rummager used the document type from SpecialistPublisher.document_types as the
type when deleting the old documents from Rummager. For ESIF documents the
type name is different in the indexable formatter from that in
SpecialistPublisher.document_types, so the old documents were not correctly
deleted from Rummager, leaving them as duplicates after the script had been
run to create the documents with the new id.

This change to the existing script will ensure that the old ESIF documents are
correctly deleted. All other formats have the same type name in both places so
ESIF is a special case here. The indexable formatter for a document type isn't
available directly from the wiring, so this is the simplest change to get the
correct type name in the script.

I've made this change to the existing script rather than writing a new one so
that it's clearer in future that the original script did the wrong thing.