Use separate services for GDS/non-GDS editors
When the User comes to add an attachment, if they were a GDS editor
editing a non GDS manual then this would come back with a "Manual not
found". This commit fixes that by using a new registry which isn't
scoped to an organisation for GDS editors.

This approach introduces some duplication. This is intended as it keeps
the attachment code similar to the code that deals with manuals. The
hope this is that this duplication can be removed more easily at a
later date.

Ticket:
https://trello.com/c/yo9nTcYO/85-attachments-can-t-be-added-to-manual-sections