Send `public_timestamp` to rummager
`last_update`  and `public_timestamp` mean the same thing in rummager.

Current behaviour of rummager is that if `public_timestamp` is empty,
`last_update` gets copied into `public_timestamp`.

This commit changes what is sent to rummager so that we populate
`public_timestamp` directly and can remove the special field and
special code from rummager.