Refactor finder publishing
To create the finder content items we were previously running a glob
over two different folders and then zip'ing the resultant file arrays
together. This assumed that both sets of files would be read of disk in
the same order.

This changes that so that instead we now read both folders but then only
select the files that exist in both folders. We then load each of those
files by name so that we can be sure that the files are matched together
correctly.

To do this I have refactored the arguments to the publishing publisher
to take a single object which is an array of finders. Each finder
contains both the schema and the metadata to then create the finder
from. Internally I have modified many of the methods to pass the whole
finder object around rather than passing just the metadata or schema as
they are individually needed.