Move finder loading logic into its own module
So that we can add tests and extra logic around returning lists of
non-matching files move the logic around finding and loading the finder
data into its own class.

This also highlighted a couple of typos and mistakes in the original
once I started adding code, which I have fixed as part of this.