Replace boolean with polymorphism
There're a few parts to this.

Boolean arguments are hard to read. If they're the only option available, they
should probably be keyword arguments so that their intention is as clear in the
method's use throughout the codebase as it is in the method's definition.

Conditional control flow has a habit of growing in complexity and becoming
unreadable. For that reason, some people prefer to avoid introducing it in the
first place. A fine example of conditional logic that's grown to be difficult
to understand is in PublishingApiFinderPublisher#call, in one of the files
affected by this commit, and some of the conditionals I've removed here already
live inside nested conditionals themselves.

This project favours the use of methods over direct access to
`@instance_variables`. I assume that's because
- on their own, even if never changed, methods're visually less noisy
- methods're easier to swap out with a different implementation, without having
  to remove all the `@`s

The Logger object is there for the using and is preferred to both
- passing boolean arguments, and
- adding conditionals
using objects like the Logger often negates the need for conditionals since the
particular flavour of Logger we send our messages to can decide how to handle
them.

A note on keyword parameters. Personally I'd say these are preferable when
- the typical values expected as arguments don't speak for themselves,
- the name of the method doesn't imply what its arguments should be,
- and also possibly when the arguments are optional
It does appear to be conventional across GOV.UK code to use a keyword for
a `logger` argument, so I've used the same style here.