Move and rename RSpec shared_examples
There are other shared_examples in this repository, but they are defined and
used within a single spec file.

I think the renaming of the files (to remove the "_spec") is the biggest
change, since that _might_ be what's silenced a bunch of RSpec warnings*.

* Those warnings looked like:
```
WARNING: Shared example group 'an indexable formatter' has been previously defined at:
  /var/govuk/specialist-publisher/spec/exporters/formatters/abstract_indexable_formatter_spec.rb:1
...and you are now defining it at:
  /var/govuk/specialist-publisher/spec/exporters/formatters/abstract_indexable_formatter_spec.rb:1
The new definition will overwrite the original one.
```

As part of this change, instead of replacing the affected specs' `require`s,
I removed them and everything worked fine. (Compared to removing the `require`s
_before_ moving/renaming the shared_examples files, which left RSpec not being
able to locate the shared_examples)