Don't index preview_only Finders on Production
In order to prevent Finders going into Elastic Search on Production we use
the `preview_only` flag to decide if it should be published or not. This
commit follows the same logic in the `PublishingApiFinderPublisher` but
keeps them separate. There's nothing to stop these being merged in
future but as this is preventing deploys I thought I'd go down the
simplest route now.

[Ticket](https://trello.com/c/Nn0DoEN2/120-preview-only-finders-should-not-be-published-to-rummager).