Delegate persisted detection to the document editions
`updated_at` doesn't seem to be set correctly on the document
instance. The document is really just a wrapper around a
collection of editions, so to check if a document is persisted,
we need to check if any of the editions are persisted.

This fixes a bug where `persisted?` was returning a false negative
when a document was invalid, causing the form to post to the
`#create` method. This caused duplicate documents to be created,
which would then be invalid due to the duplicate slug.