Validate sub-category present if parent has sub-categories
Sub-category validation is being extended based on a meeting with HMCTS.
Some parent categories will not have any sub-categories related to them.
A tribunal decision with a blank sub-category will be valid if the parent
category has no sub-categories. Otherwise presence of the sub-category is
required.

This is a solution to a concern raised from the Finder team that
it was possible for bad category/sub-category data to be entered by
admin users creating tribunal decision documents.