Send _name single value label to rummager as text not array
The values suffixed with _name are used to store facet
label text in rummager.

For _name values that are always a single label, send to
rummager as a text value instead of as a single element array.
There is no need to send them as an array.

For example tribunal decision category and sub-category are a
single value, so can be sent to rummager as text not an array.

This will be tested in feature integration tests in a following
commit.