Use `publish_pre_production_finders` config
Stop using the `PUBLISH_PRE_PRODUCTION_FINDERS` environment variable directly.

It's hard to say whether this is cleaner or just an added level of indirection,
but I'm not convinced using environment variables directly in multiple pieces
of application code is great either!

The same approach was taken
https://github.com/alphagov/hmrc-manuals-api/commit/c7c2472f1e7c564723196a641e187dc7f765f0d2,
so it's worth trying to be consistent with that.