Add tax tribunal decisions finder
Add a finder for Upper Tribunal (Tax and Chancery Chamber) decisions.
Upper Tribunal (Tax and Chancery Chamber) publishes decisions that
describe the outcome of a tribunal case.

HMCTS may decide to also publish First-tier Tribunal (Tax) decisions
using this same finder. The base_path slug has been named
to be flexible enough to allow for this.

Tribunal decision finders are used by public, professionals and the
tribunals themselves to search for past decisions.

Published decisions form a part of case law and are used for research
purposes and to prepare for current or future cases.