Add Employment Tribunal decisions finder
Add a finder for Employment Tribunal decisions.
The Employment Tribunal publishes decisions that
describe the outcome of a tribunal case.

Tribunal decision finders are used by public, professionals and the
tribunals themselves to search for past decisions.

Published decisions form a part of case law and are used for research
purposes and to prepare for current or future cases.