Add a script to report on duplicate Specialist Documents
We fixed a bug recently that was causing new document ids to be
assigned to editions for existing documents. This caused what appear to
be duplicate documents to be created, where one slug is represented by
multiple document ids. This script will report on these instances as a
CSV.

Run as `bundle exec bin/find_duplicate_documents >
duplicate_documents.csv`