Output old and new slugs on stdout for import
When the import process is run, the old documents need withdrawing
in Whitehall. In order to do this, we need to know both the old and
new URLs.

This doesn't output them in a very nice way, but it's only a small
number of documents affected, so a manual process is fine.