Add Employment Appeal Tribunal decisions features
These feature integration tests cover the same core scenarios for
Employment Appeal Tribunal decisions that have
been tested for other finders, such as the AAIB report finder.

The scenarios cover creating and editing, publishing, withdrawing
an Employment Appeal Tribunal decision, and
adding and replacing attachments to a decision.

It was easiest to add these in a single commit, as they were
developed at the same time based on a copy of the features for
asylum support tribunal decisions.