Add Documents to Rummager
This commit adds a `SearchPresenter` class to handle the logic around
presenting the Document for Search. It also adds Rummager to the
`DocumentsController` and adds the document there. Of note is the way
the `SearchPresenter` fetches the Org slug. Within Specialist Publisher
we only have the `content_id` of the Org, so it needs to fetch this
info from the Publishing API and then do 2 gsubs on it, to remove the
possible slug prefixes that the `base_path` will contain.