Add Loader and Publisher classes for Finders
This commit adds the classes to Load and Publish the Finders to
Publishing API V2 and add them to Rummager. It also adds the feature
flag for pre-production finder publishing.

This commit also changes the code to reflect the fact that the schema
and metadata files are now one file.