Temp fix for missing details
As `details` is currently not expanded from the index `GET` endpoint
this is a temp fix to request each individual payload when you hit the
index endpoint for a format. This is obviously an N+1 query but by the
time this is going into Production we should be expanding `details`.