Reject blank metadata for Pub API
When developing the App I went by what specialist documents already
existed. These documents had some metadata with blank values (`""`).
This was incorrect with what the Publishing API is meant to have but
also doesn't match the content schema for this format. This commit
removes any metadata which is empty by calling `reject` on the metadata
hash.