Add section links to manual only when it is new
This commit checks that section is not already existing in manual links
hash, and if it is existing, it will save a patch request to publishing
api.

Also added a #match_any_body which is a function that always return
true, this a work around for using gds-api-adpaters' test helpers so
that it will be agnostic to the request body. Useful for testing that a
request is NOT made.