Don't send email alert on publish for dsu format
This requirement is listed as a feature on the previous version of specialist-publisher [here](https://github.com/alphagov/specialist-publisher/blob/master/features/publishing-a-drug-safety-update.feature#L30-L60).

According to [this commit:](https://github.com/alphagov/specialist-publisher/commit/d3c7922380f17e4f6d81e538dc7082c5215363d6)

“The Drug Safety Update is a monthly newsletter sent by the MHRA with custom
content (the full content of the document body). Therefore we're going to allow
the MHRA to send this out manually rather than alerting the user each time a
new one is published.”.

Drug Safety Updates didn’t seem to be sent out from the previous version of specialist-publisher so this is possibly done through Gov Delivery.

This PR ensures that Drug Safety Updates are published but email-alerts are not triggered automatically as they are with the other formats.