(3/3) Render Govspeak - Wire up Govspeak presenter
This is the last of 3 commits, and contains the main feature changing
code. This changes what is in the body of the payload that we send to
publishing API.
By using the Govspeak presenter, we send an array which contains two
hashes, one for govspeak and one for html version of the body.

The Document#find method is compatible with single string or
multi-content-type body, the spec is only in cma_case model but I plan
on refactoring such common behaviours into a document spec file

This commit is coupled with an update on govuk-content-schema