Upgrade ruby to 2.2.4
It's good to be as up-to-date as possible with ruby upgrades to get
the latest language features and security updates.

I haven't upgraded to the latest version (2.3.0) because of deprecation
warnings that originate from `gds-api-adapters` (which in turn come from
the `rest-client` dependency):

[warning] The response contained in an RestClient::Exception is now a RestClient::Response instead of a Net::HTTPResponse, please update your code

See https://github.com/rest-client/rest-client/issues/454
for details.

The medium-to-long-term solution is to break the `gds-api-adapters`
dependency on `rest-client` and just use `net/http`.