Add first_published_at field to Drug Safety Update specialist-document
  - Drug Safety Update(DSU) specialist-documents need to be searchable by
    `first_published_at`
  - Add `first_published_at` field to DSU specialist-document
  - `first_published_at` value is set when the DSU specialist-document is first published
  - If the DSU specialist-document is re-published `first_published_at` is not updated

We add two tests, one for publishing an unpublished DSU, the other for republishing a DSU.

We update the sample spec payload for DSU with an entry for `change_history`. This allows us to simplify our tests, otherwise we have to explicitly set the `change_history` when we check our expected payload against what's being sent to the PublishingApi.

When updating `first_published_at` fails, we return out of the `publish!` method with `false`, as we do not want to continue publishing or sending to Rummager.

Note that we had change the HTTP status code check for the Rummager request to 202 as that's what it appears to be returning now. The Document Class's (which DrugSafetyUpdate inherits from) `publish!` method still relies on the Rummager request returning 200, so that might have to be updated in a future PR.