[RED TESTS]Add user authorisation to attachments_controller
Move `after_action :verify_authorized` from documents_controller to
application_controller. This allows pundit to throw a warning when we
expose a controller action without proper authorisation rules in place.

We extract ApplicationPolicy as the parent class to reduce duplicate
helper methods. However, we allow duplicate policy-rule methods so it
is easier to read all the rules for each controller in one place.

This commit is failing some tests but those shall be fix with the next
few commits when policies are implemented for other controllers.