Pass tests by adding policy for all controller actions
The policy related to manuals are completely open right now, but there
are some permissions/scoping in place such as the code in
`ManualsController#index`. We decided to first do a generic rule to
quiet down pundit's errors, and when we play the card for manuals
permissions, we can implement authorisation rules for manuals.

There was also a change in routes, because we had been quietly passing
through the `DocumentController#index` actions which redirects us back
to `ManualsController#index` when we go to root.