Unit tests for document policy
The design of this test is to focus on revealing what the 3 roles mean
in code, for example, a writer is a user without editor or gds_editor in
their permissions attribute. This is because the source code for policy
classes documented what permissions are given to which role, but it is
not obvious what is being checked when we say we are checking for a
departmental editor.