Make SearchPresenter get real orgs
A retrievable-by-author-only restriction in the publishing-api:
https://github.com/alphagov/publishing-api/blob/master/app/controllers/v2/content_items_controller.rb#L64-L68

means that `get_content_items` always returns an empty organisations
list. This means that Rummager never gets to hear about organisation
slugs and hence can't find our specialist content when including slugs
in search terms.

Daniel Roseman pointed out we should be using `get_linkables`
instead, so we stub that and add tests for everything else (there was no
 spec for SearchPresenter).

Further, Tijmen Brommet pointed out that this will not be necessary,
since 'rummager will start looking up the organisations in the
publishing-api soon-ish'. This is a good thing, as `get_linkables`
for a document type of 'organisation' is very expensive and can't be
filtered.