Remove organisation_slugs from SearchPresenter
Previously, the specs were misleading inasmuch as they were asserting
that we had sent details of organisations to Rummager. Outside of
the stubs in the specs, and in real-world use,
`publishing_api.get_content(document_type: 'organisation'...)` had
begun to return an empty list, so in reality we were never sending
organisation details to Rummager, despite the specs asserting that
we were.

This commit removes the misleading stubs and assertions about what
indexable attributes were sent to Rummager, and leaves a placeholder
for us to address the organisations issue when upstream Rummager work
 around organisation details is complete.