Remove the placeholder for orgs in SearchPresenter
* We're asserting that later, Rummager will step in and use the
indexing queue (consumed by rummager-sidekiq) as a notification list
rather than a source of truth, and will be asking the publishing-api
for truth. As a result, we won't need to do anything here (except
perhaps remove further fields to avoid giving the impression that
Rummager will continue to treat anything here as authoritative).