Add ‘updated_at’ to Payload responses
The Publishing API returns the updated_at field
which is used in the view, so we need to include
this in the response stub.

There’s some tech debt because the Payloads are
also used for setting up Publishing API request
assertions and the request/response payloads are
different. We’ve written up a ticket here to
address this:
https://trello.com/c/4ZLEWBdg/136-clean-up-payloads

For now, we’ve deleted the payload_spec.rb as it
is misleading – i.e. it’s validating a response
payload against a request schema.