Update finder org access to be more like SPv1
In this app, the organisation lists in the finder schemas determine which publishers (from which org)
can see and edit particular finders.

In Specialist Publisher v1, this info is not in 1, but two places:

- the [finder metadata files](https://github.com/alphagov/specialist-publisher/tree/master/finders/metadata)
- the [permission checker](https://github.com/alphagov/specialist-publisher/blob/master/app/lib/permission_checker.rb)
which actually controls whether publishers from specific orgs can edit those document types

In V1, the organisation lists in those two places went out of sync with time.
When finders were migrated across to this app, the (outdated) metadata orgs from V1 were
used to create the schemas. Instead, the orgs in the permission checker are really the source of truth
for access.

This commit makes brings the org lists in line with the permission checker in V1.