Clean up usage of public_updated_at
Since we are now not writing public_updated_at to publishing-api,
there's no reason to update it on save.
That removal is coupled with the usage of the same timestamp in change
history.

As we have now found that the timestamp is currently incorrectly set in
change-history, I took the liberty to flag this up by using an ugly
constant STUB_TIME_STAMP. This should mark all the places that needs
updating when we play future stories relating to that feature.