Remove headers from document factory defaults
Adding headers to govspeak requires the inclusion of a separate headers section in the payload (code for this lies in DocumentPresenter). By adding this to the default DocumentFactory (and all factories that inherit from it), all tests implicitly test that our header logic works, which may be confusing when testing documents with no headers. It also may require to delete the 'headers' section from a payload for individual test. This commit aims to alleviate this by removing headers from the default body text.