Implement attachment snippet expansion
When converting the document body from Govspeak to HTML, we also expand any attachment snippets to functioning links. The resulting HTML is placed into the 'test/html' content type for the body section in the payload that will be sent to the publishing api.

We refactor the `GovspeakPresenter` to be a standard class that initialises an instance to perform its work. This allows for use of instance variables and instance methods and makes the code easier to read. We also rewrite the specs for the new interface.

Additionally we add to the feature tests for editing a document and adding attachment snippets to the body. Note that in order to test that the snippet is still in its unexpanded form after clicking on 'Save' we have to restub the payload that the publishing api sends back with the updated payload. It's probably not ideal to assert on that, but we do assert that the same content has been sent to the publishing api, so might be fine...