Move updated_at document test to attachment presenter test
The test is exercising the logic in AttachmentPresenter. It was already tested indirectly in the "renders the correct data" test in attachment_presenter_spec.rb, but we create a new test to draw attention to the fact that `updated_at` will get updated to the current time whenever the `present` method is called. We remove the redundant test from document_spec.rb.