Fix a Rails autoload circular dependency problem
For reasons unknown, Rails was raising this error on integration:
RuntimeError: Circular dependency detected while autoloading constant Services

Add explicit 'require' statements fixes this problem. I'm not sure
why this was happening as 'lib' was on the load path and there's only
one 'Services' in the app. Maybe it was clashing with something inside
Rails.