Always extract change_history in from_publishing_api
Prior to this change we would only extract the change_history if the
document was what we considered "published" (e.g. "live" or "rerafted").
If we don't extract the change_history from the API payload then when
we write the document back to the publishing-api we will clear out the
change_history.  This feels overly destructive and something that we
should take a specific and explicit step to achieve if we do want it
rather than the current approach which feels like it might be an
oversight.