Add Document#has_ever_been_published?
The current implementation is to look for the presence of the "First
published." note that we add during publish! - if we find it then we
assume the document has been published before, if we don't then we
assume it hasn't.  We then use this new method to decide if we should add
the "First published." note during the call to publish!.

This isn't the best way to decide if a document has ever been published
but the alternatives all have problems.  first_published_at is one
candidate as it is added by publishing-api on first publication, but some
formats (e.g. dfid research output) allow editing this field before
publishing so we can't rely on it.  We also can't use the state of the
document because that can only tell us if the document is currently
published not if it has ever been published (for example a document can
go straight from draft to unpublished without going through published
beforehand).

There is a story to play soon that will add full state_history to the
response from the publishing-api at which point we'll update the
implementation of has_ever_been_published? to use this information and
be more robust.