Add published trait for document factories
Our understanding of what a "published" document is has changed a bit
now that we consider the contents of its change_history as marker. In
order that our tests are more closely aligned with real data we introduce
a published trait for the document factory that makes sure the expected
entires are in the change_history.  We also make change_history an
attribute that can be specified without digging deep into the details
hash.

We've updated some of the specs to use this new trait.