Do not republish things that can cause side-effects…
If we were to republish something in an
‘unpublished’ state, it exposes a risk that it
could be made accidentally live if a subsequent
unpublished request fails. There is no way to
ensure that these operations happen as a single
atomic operation in the Publishing API.

Our best option in this case is to skip the
republishing of this document. It may be the case
that whoever runs the republishing script needs
to do something to account for those documents
that have been skipped.

The RepublishWorker will skip over any other
states that it doesn’t know about (or that may be
introduced in the future) as it doesn’t know the
correct set of requests to issue to ensure the
document is republished into that state.