Add guard clause where we call time_ago_in_words
As `last_edited_at` field is rather new from publishing-api, we will
need to do some data migration to populate that field for existing
documents. In the meantime, a nil value will cause an error at
`time_ago_in_words` so I'm adding a guard clause for a smoother
transition.