Consolidate and fix ‘actions’ behaviour…
The logic that governed which buttons were visible
and the text that displays alongside those buttons
was scattered throughout the app. This commit
brings it all together into an ActionsPresenter
that can be easily tested for the various
combinations of states.

This commit also changes the messaging for some
state combinations that were not accounted for
previously. Specifically, the ‘unpublished’ state
didn’t have much display logic around it when it
should have.