Preparatory changes for removing redrafted
Changed the way first_draft operates so that it uses first_published_at instead of the state history to check for drafts.
However, since DFID documents can have draft documents with first_published_at we must determine first_draft? using
state_history for DFID, as was previously used for all formats. This decision was made due to the desire to reduce
reliance on state_history in the back end.

We condensed a lot of logic regarding the setting of update type, such that most of it is handled when first receiving
the object, rather than both when receiving and saving, and handling. We also replaced 'search' with 'rummager' when
using Plek